refactor: remove unused persistence backend factory
